+++ python/fenics/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+"""Main module for FEniCS.
+
+This module is identical with the main DOLFIN module.
+"""
+
+from dolfin import *